Version 2
This is an updated version of WorldMaker9000. It now creates and paints a 10*10(variable) block of colors based on a 2D Boolean array that states whether each block is 'full'm and then f so, assigns it a random color. The layout has also been moved around.
The next update will mostly, if not all, be based around building a correct javadoc, which I plan to keep on top of.